requirements.txt
azure-functions==1.5.0
